[handleFileSelection func] internal: refactor
