+++ examples/main/shared/utils/mock.js
