+++ examples/main/shared/utils/mockEventsData.js
