+++ examples/main/shared/utils/mockProductsData.js
