+++ examples/main/shared/utils/verticalEmbedUtil.js
