+++ examples/storybook/stories/Components/EditorWrapper.js
@@ -31,6 +31,7 @@ import { mockFetchUrlPreviewData } from '../../../main/shared/utils/linkPreviewU
 import { pluginTextColor, pluginTextHighlight } from 'wix-rich-content-plugin-text-color';
 import MobileDetect from 'mobile-detect';
 import '../styles.global.scss';
+import { mockFileUploadFunc } from '../../../main/shared/utils/fileUploadUtil';
 
 const { Instagram, Twitter, YouTube, TikTok } = LinkPreviewProviders;
 const { event, booking, product } = verticalEmbedProviders;
@@ -53,23 +54,7 @@ const onFilesChange = (files, updateEntity) => {
 const configs = {
   fileUpload: {
     accept: '*',
-    handleFileSelection: updateEntity => {
-      const filenames = ['image.jpg', 'document.pdf', 'music.mp3'];
-      const multiple = false;
-      const count = multiple ? [1, 2, 3] : [1];
-      const data = [];
-      count.forEach(() => {
-        const name = filenames[Math.floor(Math.random() * filenames.length)];
-        const filenameParts = name.split('.');
-        const type = filenameParts[filenameParts.length - 1];
-        data.push({
-          name,
-          type,
-          url: 'http://file-examples.com/wp-content/uploads/2017/10/file-sample_150kB.pdf',
-        });
-      });
-      setTimeout(() => updateEntity({ data }), 500);
-    },
+    handleFileSelection: mockFileUploadFunc,
   },
   giphy: {
     giphySdkApiKey: process.env.GIPHY_API_KEY || 'HXSsAGVNzjeUjhKfhhD9noF8sIbpYDsV',
