+++ examples/storybook/stories/Plugins/VerticalEmbed/VerticalEmbedEditor.js
@@ -5,7 +5,7 @@ import {
   verticalEmbedProviders,
 } from 'wix-rich-content-plugin-vertical-embed';
 import PropTypes from 'prop-types';
-import MockVerticalSearchModule from '../../../../main/shared/editor/Utils/verticalEmbedUtil';
+import MockVerticalSearchModule from '../../../../main/shared/utils/verticalEmbedUtil';
 
 const { event, booking, product } = verticalEmbedProviders;
